# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API REST de videojuegos
  description: API REST para consultar y gestionar videojuegos, desarrolladores, reviews y países relacionados con sus desarrolladores.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Servidor local de desarrollo
...
paths:
...
  /games:
    get:
      tags:
        - Games
      summary: Obtener videojuegos
      description: Lista los videojuegos almacenados en MongoDB. Permite buscar, filtrar, ordenar y paginar resultados.
      parameters:
        - name: id
          in: query
          description: ID numérico del videojuego.
          schema:
            type: integer
            minimum: 1
          example: 3328
        - name: search
          in: query
          description: Texto a buscar en nombre o slug.
          schema:
            type: string
          example: the witcher
        - name: platform
          in: query
          description: Plataforma del videojuego.
          schema:
            type: string
          example: PC
        - name: genre
          in: query
          description: Género del videojuego.
          schema:
            type: string
          example: Action
        - name: store
          in: query
          description: Tienda en la que está disponible.
          schema:
            type: string
          example: Steam
        - name: minRating
          in: query
          description: Valoración mínima.
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 5
          example: 4.5
        - name: page
          in: query
          description: Número de página.
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 2
        - name: limit
          in: query
          description: Número máximo de resultados.
          schema:
            type: integer
            minimum: 1
            default: 1000
          example: 20
        - name: sort
          in: query
          description: "Campo de ordenación. Con guion inicial ordena de forma descendente. Campos permitidos: id, name, released, rating, metacritic, playtime."
          schema:
            type: string
          examples:
            ratingAsc:
              value: rating
            nameDesc:
              value: -name
      responses:
        "200":
          description: Lista de videojuegos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GamesListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - Games
      summary: Crear videojuego
      description: Inserta un videojuego en la colección principal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Videojuego creado correctamente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedResponse"
        "400":
          description: Faltan campos obligatorios o hay arrays vacíos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Ya existe un videojuego con ese ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /games/{id}:
    get:
      tags:
        - Games
      summary: Obtener videojuego por ID
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: Videojuego encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags:
        - Games
      summary: Actualizar videojuego
      description: Actualiza los campos admitidos de un videojuego existente.
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameUpdate"
      responses:
        "200":
          description: Videojuego actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateGameResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: El cambio rompería recursos relacionados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags:
        - Games
      summary: Eliminar videojuego
      description: Elimina un videojuego de la colección principal.
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: Videojuego eliminado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: El videojuego tiene recursos relacionados y no puede eliminarse.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /games/{id}/country:
    get:
      tags:
        - Games
        - Countries
      summary: Obtener países relacionados con un videojuego
      description: Devuelve en XML los países asociados a los desarrolladores de un videojuego.
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: Información de países en XML.
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/CountriesXml"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /games/{id}/reviews:
    get:
      tags:
        - Games
        - Reviews
      summary: Obtener reviews de un videojuego
      description: Devuelve las reviews asociadas al videojuego indicado.
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: Reviews asociadas al videojuego.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameReviewsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /games/{id}/enriched:
    get:
      tags:
        - Games
      summary: Obtener videojuego enriquecido
      description: Devuelve el videojuego junto con sus reviews y los países asociados a sus desarrolladores.
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: Videojuego con datos relacionados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrichedGameResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /reviews/game/{gameId}:
    get:
      tags:
        - Reviews
      summary: Obtener reviews por videojuego
      description: Recupera las reviews vinculadas al videojuego indicado.
      parameters:
        - name: gameId
          in: path
          required: true
          description: ID numérico del videojuego.
          schema:
            type: integer
            minimum: 1
          example: 3328
      responses:
        "200":
          description: Reviews del videojuego.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewsByGameResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    GameId:
      name: id
      in: path
      required: true
      description: ID numérico del videojuego.
      schema:
        type: integer
        minimum: 1
      example: 3328
  responses:
    BadRequest:
      description: Petición incorrecta.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Error interno del servidor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Game:
      type: object
      properties:
        id:
          type: integer
          example: 3328
        slug:
          type: string
          example: the-witcher-3-wild-hunt
        name:
          type: string
          example: "The Witcher 3: Wild Hunt"
        released:
          type: string
          nullable: true
          example: "2015-05-18"
        background_image:
          type: string
          nullable: true
          example: https://media.rawg.io/media/games/618/618c2031a07bbff6b4f611f10b6bcdbc.jpg
        rating:
          type: number
          format: float
          nullable: true
          example: 4.64
        metacritic:
          type: integer
          nullable: true
          example: 92
        playtime:
          type: integer
          nullable: true
          example: 43
        platforms:
          type: array
          items:
            type: string
          example:
            - PC
            - PlayStation 4
        genres:
          type: array
          items:
            type: string
          example:
            - Action
            - RPG
        stores:
          type: array
          items:
            type: string
          example:
            - Steam
            - GOG
        esrb_rating:
          type: string
          nullable: true
          example: Mature
    GameInput:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - platforms
        - genres
        - stores
      properties:
        id:
          type: integer
          example: 999999
        slug:
          type: string
          example: example-game
        name:
          type: string
          example: Example Game
        released:
          type: string
          example: "2026-01-01"
        background_image:
          type: string
          nullable: true
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        metacritic:
          type: integer
          nullable: true
          minimum: 0
          maximum: 100
        playtime:
          type: integer
          minimum: 0
        platforms:
          type: array
          minItems: 1
          items:
            type: string
        genres:
          type: array
          minItems: 1
          items:
            type: string
        stores:
          type: array
          minItems: 1
          items:
            type: string
        esrb_rating:
          type: string
          nullable: true
    GameUpdate:
      type: object
      description: Campos admitidos en la actualización de un videojuego. No se actualiza el ID.
      properties:
        slug:
          type: string
        name:
          type: string
        released:
          type: string
        background_image:
          type: string
          nullable: true
        rating:
          type: number
          format: float
        metacritic:
          type: integer
          nullable: true
        playtime:
          type: integer
        platforms:
          type: array
          items:
            type: string
        genres:
          type: array
          items:
            type: string
        stores:
          type: array
          items:
            type: string
        esrb_rating:
          type: string
          nullable: true
        developers:
          type: array
          items:
            type: string
...
    CountriesXml:
      type: object
      xml:
        name: countries
      properties:
        country:
          type: array
          xml:
            wrapped: false
            name: country
          items:
            $ref: "#/components/schemas/Country"
    GamesListResponse:
      type: object
      properties:
        total:
          type: integer
          example: 1000
        videogames_length:
          type: integer
          example: 2
        videogames:
          type: array
          items:
            $ref: "#/components/schemas/Game"
...
    ReviewsByGameResponse:
      type: object
      properties:
        reviews_length:
          type: integer
          example: 3
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
    GameReviewsResponse:
      type: object
      properties:
        gameId:
          type: integer
          example: 3328
        gameName:
          type: string
          example: "The Witcher 3: Wild Hunt"
        reviews_length:
          type: integer
          example: 3
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
    EnrichedGameResponse:
      type: object
      properties:
        game:
          $ref: "#/components/schemas/Game"
        wikidata:
          type: object
          properties:
            countries:
              type: array
              items:
                $ref: "#/components/schemas/Country"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
    CreatedResponse:
      type: object
      properties:
        message:
          type: string
          example: Videojuego creado correctamente
        id:
          type: integer
          example: 999999
        game:
          $ref: "#/components/schemas/Game"
...
    UpdateGameResponse:
      type: object
      properties:
        message:
          type: string
          example: Videojuego actualizado correctamente
        id:
          type: integer
          example: 3328
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: El id debe ser un número positivo
        error:
          type: object
          description: Detalle técnico cuando el error procede de una excepción.
          additionalProperties: true
          nullable: true
        missingFields:
          type: array
          description: Campos obligatorios que faltan en la petición.
          items:
            type: string
        invalidFields:
          type: array
          description: Campos enviados que la ruta no admite.
          items:
            type: string